+++ AMS/Models/Case.cs
@@ -4,7 +4,6 @@ using AMS.Models;
 public class Case: BaseEntity
 {
 
-
 public string CaseNumber { get; set; }
 public string CaseTitle { get; set; }
 public string CaseType { get; set; }
@@ -13,10 +12,10 @@ public string Description { get; set; }
 public DateTime FilingDate { get; set; }
 public string Status { get; set; }
 
-public int AdvocateId { get; set; }
+public long AdvocateId { get; set; }
 public Advocate Advocate {get; set; }
 
-public int ClientId { get; set; }
+public long ClientId { get; set; }
 public Client Client { get; set; }
 
 public ICollection<Hearing> Hearings { get; set; }
